[Delete] #200 - 카테고리 탭 관련 삭제
