--- CERTI-iOS/Presentation/Category/Coordinator/CategoryCoordinator.swift
@@ -1,37 +1,0 @@
-//
-//  CategoryCoordinator.swift
-//  CERTI-iOS
-//
-//  Created by OneTen on 6/25/25.
-//
-
-import SwiftUI
-
-enum CategoryRoute: Hashable {
-    case search
-    case certificateDetail
-}
-
-final class CategoryCoordinator: ObservableObject {
-    
-    //MARK: - Property Wrappers
-    
-    @Published var path = NavigationPath()
-    
-    //MARK: - Method
-    
-    //다음에 보여질 view를 navigationStack에 push
-    func push(next route: CategoryRoute) {
-        path.append(route)
-    }
-    
-    //현재 view를 navigationStack에서 pop
-    func pop() {
-        path.removeLast()
-    }
-    
-    //맨 처음으로 돌아감(navigationStack 초기화)
-    func reset() {
-        path = NavigationPath()
-    }
-}
